Python/kafka/requirements.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio